fixed Ollama post processing
The program was looking for an API key that Ollama doesn't have since it's local. I fixed it. I have no experience writing Python code, so this might break something else but for the time being it fixes the issue I had.

(For example, this description was spellchecked by `airat/karen-the-editor-v2-strict:latest`!)